+++ app/src/main/java/org/sopt/and/ui/home/component/SubVideoContent.kt
@@ -1,27 +1,26 @@
-package org.sopt.and.ui.hoem.component
+package org.sopt.and.ui.home.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowForwardIos
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -29,14 +28,14 @@ import androidx.compose.ui.unit.sp
 import org.sopt.and.R
 
 @Composable
-fun Top20VideoContent(
+fun SubVideoContent(
     contentTitle: String,
     imageList: List<Int>,
     modifier: Modifier = Modifier
 ) {
     Column(
         verticalArrangement = Arrangement.spacedBy(5.dp),
-        modifier = Modifier.padding(bottom = 10.dp)
+        modifier = Modifier.padding(bottom = 20.dp)
     ) {
         Row(
             horizontalArrangement = Arrangement.SpaceBetween,
@@ -50,38 +49,26 @@ fun Top20VideoContent(
                 fontWeight = FontWeight.Bold,
                 color = Color.White
             )
+            Icon(
+                imageVector = Icons.AutoMirrored.Filled.ArrowForwardIos,
+                contentDescription = Icons.AutoMirrored.Filled.ArrowForwardIos.name,
+                tint = Color.Gray
+            )
         }
         LazyRow(
             contentPadding = PaddingValues(horizontal = 10.dp),
-            modifier = Modifier.height(250.dp)
+            modifier = Modifier.height(150.dp)
         ) {
             items(imageList.size) { index ->
-                Box(
-                    modifier = Modifier.fillMaxSize(),
-                    contentAlignment = Alignment.BottomStart
+                Surface(
+                    modifier = Modifier.padding(horizontal = 5.dp),
+                    shape = RoundedCornerShape(10.dp)
                 ) {
-                    Column {
-                        Surface(
-                            shape = RoundedCornerShape(10.dp),
-                            modifier = Modifier
-                                .weight(1f)
-                                .padding(horizontal = 5.dp)
-                        ) {
-                            Image(
-                                painter = painterResource(imageList[index]),
-                                contentDescription = "",
-                                contentScale = ContentScale.Crop
-                            )
-                        }
-                        Spacer(modifier = Modifier.weight(0.15f))
-                    }
-                    Text(
-                        text = "${index+1}",
-                        modifier = Modifier.padding(start = 5.dp),
-                        fontSize = 60.sp,
-                        fontStyle = FontStyle.Italic,
-                        fontWeight = FontWeight.ExtraBold,
-                        color = Color.White
+                    Image(
+                        painter = painterResource(imageList[index]),
+                        contentDescription = "",
+                        modifier = Modifier.fillMaxSize(),
+                        contentScale = ContentScale.Crop
                     )
                 }
             }
@@ -91,7 +78,7 @@ fun Top20VideoContent(
 
 @Preview
 @Composable
-private fun Preview() {
+private fun PreviewVideoContent() {
     val imageList = listOf(
         R.drawable.image1,
         R.drawable.image2,
@@ -99,23 +86,10 @@ private fun Preview() {
         R.drawable.image4,
         R.drawable.image5,
         R.drawable.image6,
-        R.drawable.image1,
-        R.drawable.image2,
-        R.drawable.image3,
-        R.drawable.image4,
-        R.drawable.image5,
-        R.drawable.image6,
-        R.drawable.image1,
-        R.drawable.image2,
-        R.drawable.image3,
-        R.drawable.image4,
-        R.drawable.image5,
-        R.drawable.image6,
-        R.drawable.image1,
-        R.drawable.image2,
     )
-    Top20VideoContent(
-        contentTitle = "오늘의 Top 20",
+
+    SubVideoContent(
+        contentTitle = "믿고 보는 웨이브 에디터 추천작",
         imageList = imageList
     )
 }
