+++ app/src/main/java/org/sopt/and/ui/home/navigation/HomeNavGraph.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.ui.hoem.navigation
+package org.sopt.and.ui.home.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavGraphBuilder
@@ -6,7 +6,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import org.sopt.and.core.navigation.MainTabRoute
-import org.sopt.and.ui.hoem.HomeScreen
+import org.sopt.and.ui.home.HomeScreen
 
 fun NavHostController.navigateToHome(navOptions: NavOptions) =
     navigate(MainTabRoute.Home, navOptions)
