+++ app/src/main/java/org/sopt/and/ui/main/MainScreen.kt
@@ -7,23 +7,16 @@ import androidx.compose.animation.slideInVertically
 import androidx.compose.animation.slideOutVertically
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.ScaffoldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import org.sopt.and.core.navigation.MainTabRoute
-import org.sopt.and.ui.hoem.HomeScreen
-import org.sopt.and.ui.hoem.navigation.homeNavGraph
+import org.sopt.and.ui.home.navigation.homeNavGraph
 import org.sopt.and.ui.main.component.MainBottomTabsBar
-import org.sopt.and.ui.my.MyScreen
 import org.sopt.and.ui.my.navigation.myNavGraph
-import org.sopt.and.ui.search.SearchScreen
 import org.sopt.and.ui.search.navigation.searchNavGraph
 import org.sopt.and.ui.signin.navigation.signInNavGraph
 import org.sopt.and.ui.signup.navigation.signUpNavGraph
